CROP-RECOMMENDATION/requirements.txt
numpy
streamlit
seaborn
pandas
matplotlib
scikit-learn
pillow
